third_party/WebKit/LayoutTests/fast/multicol/insane-column-gap.html

<!DOCTYPE html>
<div id="container" style="width:500px;">
    <div data-expected-width="500" style="width:fit-content; column-gap:20000000px; columns:1000 10px;"></div>
</div>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/check-layout-th.js"></script>
<script>
    checkLayout("#container");
</script>
